Catch not enough flux voltage points for fit
ReparkingRamsey now raises an exception in case only two
flux voltages are to be measured.

For the parabola fit in the analysis at least three points
are necessary to get a meaningful fit. In addition the method
used to distinguish USS and LSS does not work with only two
points.