multi_sweep_function: add methods add_sweep_function(s)
The new method adds the passed sweep_functions to the multi_sweep.
The sweep_functions can also be qcodes.Parameter objects in which case
they are converted to sweep_functions before being added to
self.sweep_functions. The methods also take care of changing the unit
of self and call check_units to make sure all sweep_fucntions are compatible.