Revert "enforce single element respects acq element assignments."
This partially reverts commit 43124761eb10536002a147f6ee2cd31361b031e1.

Reverted because this leads to a bug with compensation elements not
being part of the ese segment.

ese should anyways not be used on acquisition devices because it would
result in not triggering the acquisition at the right moment.