acr: Add function to rearrange steps in parallel groups
Not necessary anymore to copy and paste the whole code in each routine, just need to call the function split_step_for_parallel_groups. Also the labels of the split steps is now the same, so that the label-specific settings can be used. Previously, the labels where step_label_qb1qb2, which made it ambiguous to have label-specific settings. Qubit-specific settings should be specified using the qubits keyword.