acr: Fix SingleQubitCalib routine template creation
The settings are now retrieved correctly. It is possible to specify label-specific settings, including the transition name. For instance, it is possible to specifiy the settings in 'ramsey_large_AD', 'ramsey_large_AD_ge', and 'ramsey_large_AD_ef'. The settings that specify a transition will have priority over the settings that do not.