CRITICAL: update `fitting_models` module such that Qubit_dac_to_freq_precise returns float.
This function is used to calculate the frequency of a qubit with the 'transmon' model (i.e. without considering the coupling to the resonator). It is now updated to return a float and not array([float]), similar to the more `Qubit_dac_to_freq_res` function which deals with the model 'transmon_res'.