qe res spec: bugfix SHF start using enforce_pulsar_restart
With the new logic introduced in !219 the df (int_avg_det_spec) will get None as AWG
in case of a hard res spec on a SHF QA channel. This causes the
QA channel device not to get started. This bugfix ensures that pulsar
is set as the AWG in this case and therefore all involved AWGs are
restarted. FIXME added to think about a solution to only restart devices
that actually need to be restarted (QA channel device in this case).