ana therm pop: do not plot cal points for unprepared ef rabi
For an ef Rabi only on the thermal population, the Rabi oscillation amplitude
is very small. To ensure that the fit quality can be verified from the analysis plots
we decided no to plot the calib points as this fixes the plot ylim to [0, 1]. For the
prepared ef Rabi, we still plot the calib points for verification.