qe residual zz: initial commit of new qe and corresponding analysis
Measures the residual ZZ coupling by measuring two Ramsey experiments
on a qubit, one with the control qubit in the ground state and one where it is
excited before the Ramsey sequence. Comparing the frequency of the Ramsey
experiments on ecan extract the residual ZZ coupling between the two qbs.