qe residual zz: add inter-task compatibility check
We overwrite preprocess_task_list to check that the qubits used in
different tasks do not overlap. This ensures that the measurement
can be run but does not ensure validity of the results. We therefore
warn the user as soon as more than one task exist in the
preprocessed_task_list. More details in comments of the
impemented method or the class docstring.