Avoid to apply unintentionally mirror pattern to the drive readout pulse
- Current implementation of flux pulse assisted readout is a bit hacky because
the flux pulse is an attribute of the readout (drive) pulse,
and therefore segment.resolve_mirror() cannot easily detect the mirror pattern
on the child flux pulse of the readout. The existing code made use of a mirror_pattern
in the drive pulse to activate the mirror pattern of the child pulse but resulted
to the drive pulse being mirrored if there was a non trivial mirror pattern in
ro_flux_mirror_pattern but no child flux pulse (e.g. because ro_pulse_type is a
standard GaussFilteredCosIQPulse).

- This commit fixes this bug by removing the mirror pattern in the drive pulse
and checking for a child flux pulse in case no mirror pattern is found for the parent
pulse.