Segment: automatically group RO pulses into appropriate acquisition elements
The old code has led to unexpected behavior when adding RO pulse for N
measurement objects without specifying element names: each RO pulse
was then put into a separate element. Other parts of pycqed would then
assume that there are N acquisitions on each of the N mobjs.

The new code allows RO pulses for different mobjs to be in the same
element, and puts the pulse for a mobj into the first acquisition
element not yet used for this mobj.