Add possibility to enforce adding a final readout in parallel_sweep
- When using Multitasking experiment, ro_qubits is overwritten by using the values
in meas_obj_names, which itself is constructed based on ro_qubits passed by the user
and in the task_list. However, it adds only a readout for pulses which don't have a readout
in the parallel_block functions. This commit allows to bypass this automated detection
and enforce to have a final readout for all ro_qubits independently of whether or not
they have already a readout pulse in the block function.
This is useful e.g. when characterizing a reset block which contains readouts already
and a final readout is needed for the characterization of the last round of feedback based reset.