hlp_mod: bugfix: hlp_mod.get_param_from_group() crashed because it needs 'not found' item in data dict of get_params_from_files()
If a parameter is not found in the files, 'not found' will be added to the data dict. Before, the parameter was not added at all but this caused problems with hlp_mod.get_param_from_group(), since this function relies on the logic of the tag string 'not found'.