Remove return type for analysis_daemon.execute_job().
Using this particular return type causes circular import with base_analysis.
In particular because get_analysis_object_from_hdf5_file_path() references
analysis_daemon. The only way to solve this without refactoring the code
would be reloading the base_analysis in analysis_daemon (as it was before).
Howerver, this potentially introduces runtime changes in behaviour that
might be unwanted.