tda: do not crash in _get_default_data_to_fit in some cases with rotate=False
The prior implementation crashed in cases where cal_states_rotations
was empty, which is in particular the case if rotate is set to False
via the options dict and cal states were measured, but
cal_states_rotations was not provided explicitly.

In this case, having data_to_fit empty is a reasonable fallback because
only raw data gets plotted in this case.