mixer: set ro_fixed_lo_freq to shf-compatible value
In the temporary values used within the mixer calibration functions,
set the ro_fixed_lo_freq to `{"step": 100e6, "min": 1e9, "max": 8e9}`
instead of None so that this also works with SHF acquisition devices.
Make sure to set this first in the temporary values prior to other
readout values so that they are updated based on this.