Improve compatibility with python 3.12
In python 3.6, incorrectly formatted string escape sequences (\x, where
x is some character other than n, r, or t) were deprecated.
DeprecationWarnings are typically silenced in python interpreters, so
these warnings went unheeded.

Python 3.12 makes these a syntax warning, so we now see a bunch of these
if we attempt to use pycqed on python 3.12. The solution is to either
turn the string into a raw string (prefix with 'r') or add an extra
backslash to escape the existing slash.

This commit uses ruff check . --select W605 --fix to automatically
correct the 260-odd such cases that we had in pycqed.