analysis_v3: Set axes-bg transparency to be False by default
The processing pipeline's plot function assumes that the plot generated
is for a presentation with a non-white background. We enhance the
pipeline following its design transporting information using
dictionaries to enhance and default (to False) a `transparent_bg` key in
the plot_dict.

The new default is non-transparent background around axes.