SHFQA/QC/SG: set default grid of center frequency to 200MHz
This grid is used, in particular, for spectroscopy-type experiments.
For other experiments, users need to adjust ro_fixed_lo_freq
in the qubit objects.

Using a 200 MHz grid is mandatory for firmware versions 24.10 and newer.
In older versions, it is recommended for the following reason.

For recent SHF devices with LPN synthesizer, the phase-frequency
detector (PFD) in the PLL of the LPN works at 200 MHz in order
to minimize phase noise. The output channels are not phase-aligned
if the output frequency is not a multiple of 200 MHz (if multiple of
100 MHz, a 180 degree phase shift between channels is possible).